put primitive method in quantum op to inherits in parametric
